Refactor Question model to use userName instead of author
This commit updates the Question model in the 'entities' package to remove the 'author' attribute and replace it with a 'userName' attribute. The 'author' attribute was causing issues during testing because it could not be left blank, and it was difficult to create multiple users for testing purposes. The 'userName' attribute is easier to work with and provides the necessary information for displaying the name of the user who posted the question.

The 'userName' attribute is a String type and is set to the first name generated by the Faker library in the Application class. The QuestionRepository is updated to save the 'userName' attribute along with the other question details.

The following changes have been made to the Question model:
- Removed 'author' attribute
- Added 'userName' attribute of type String

The following changes have been made to the Application class:
- Updated commandLineRunner() method to generate fake data with 'userName' attribute instead of 'author'

These changes are temporary and are only intended to facilitate testing. Once testing is complete, the 'userName' attribute can be removed and the 'author' attribute can be reinstated. This commit serves as a temporary workaround until a more permanent solution is implemented.