Refactor pagination and user model addition in QuestionController
This commit refactors the pagination logic in the QuestionController by moving it to a separate private method called paginate. It also moves the logic for adding the list of all users to the model to a separate private method called addUsersToModel. The refactored code is now cleaner and more readable.

Additionally, the isTag and isSearch model attributes have been set to false in the questions method to indicate that it is just listing all sorted questions and not filtering by tags or search keywords.