Refactor SearchQuestion method in QuestionController
This commit modifies the SearchQuestion method in the QuestionController by adding a POST and GET mapping to the RequestMapping annotation to indicate that the method now handles both types of requests on the /search path. Additionally, the method now uses the addUsersToModel private method to add the list of all users to the model.

The logic for sorting questions and searching for keywords has also been refactored to use a new list called allSortedQuestions to store all sorted questions and then remove duplicates when searching for keywords. The isTag and isSearch model attributes have been set to true and false, respectively, to indicate that the search is being performed. The keyword is also added to the model as a new attribute.

As a result, this modification improves the code structure and makes the SearchQuestion method more robust and versatile by handling both POST and GET requests and removing duplicates when searching for keywords.